test: add tests for new classes > remove old tests referencing decoupled functionality > add sanitizer tests
