+++ test/src/unit_tests/mixins/validate_args_mixin_test.dart
@@ -1,161 +1,61 @@
+import 'package:collection/collection.dart';
 import 'package:magical_version_bump/src/core/mixins/command_mixins.dart';
-import 'package:mason_logger/mason_logger.dart';
-import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 class _FakeArgsValidator with ValidatePreppedArgs {}
 
-class _MockLogger extends Mock implements Logger {}
-
 void main() {
   late _FakeArgsValidator validator;
-  late Logger logger;
+  late ListEquality<String> listEquality;
 
   setUp(() {
     validator = _FakeArgsValidator();
-    logger = _MockLogger();
+    listEquality = const ListEquality<String>();
   });
 
-  group('basic errors with no set path', () {
-    test('returns error when no args are passed in modify command', () async {
-      const baseError = 'Missing arguments';
-      const verboseError = 'Additional arguments for this command are missing';
+  group('validates and returns base error', () {
+    test('returns undefined flags', () {
+      final args = ['bump', 'undefined'];
 
-      final validated = await validator.validateArgs(
-        [],
-        isModify: true,
-        userSetPath: false,
-        logger: logger,
-      );
+      final undefinedFlags = validator.checkForUndefinedFlags(args);
 
-      expect(validated.invalidReason, isNotNull);
-      expect(validated.invalidReason!.key, baseError);
-      expect(validated.invalidReason!.value, verboseError);
+      expect(listEquality.equals(['undefined'], undefinedFlags), true);
     });
 
-    test('returns error when undefined flags are passed', () async {
-      const undefinedArg = ['undefined-arg'];
-      const baseError = 'Invalid arguments';
-      const verboseError = 'undefined-arg is not a defined flag';
-
-      final validated = await validator.validateArgs(
-        undefinedArg,
-        userSetPath: false,
-        logger: logger,
-      );
-
-      expect(validated.invalidReason, isNotNull);
-      expect(validated.invalidReason!.key, baseError);
-      expect(validated.invalidReason!.value, verboseError);
-    });
-
-    test('returns error when flags are duplicated', () async {
-      const duplicatedArgs = ['major', 'major', 'name', 'yaml-version'];
-      const baseError = 'Duplicate flags';
-      const verboseError = 'Found repeated flags:\nmajor -> 2\n';
-
-      final validated = await validator.validateArgs(
-        duplicatedArgs,
-        userSetPath: false,
-        logger: logger,
-      );
-
-      expect(validated.invalidReason, isNotNull);
-      expect(validated.invalidReason!.key, baseError);
-      expect(validated.invalidReason!.value, verboseError);
-    });
-  });
-
-  group('modify command errors', () {
-    test('returns error when action flag is not first', () async {
+    test('returns error when action flag is not first', () {
       const args = ['major', 'bump'];
-      const baseError = 'Wrong flag sequence';
-      final verboseError =
-          "${validator.actions.join(', ')} flags should come first";
+      final error = "${validator.actions.join(', ')} flags should come first";
 
-      final validated = await validator.validateArgs(
-        args,
-        isModify: true,
-        userSetPath: false,
-        logger: logger,
-      );
+      final returnedError = validator.checkModifyFlags(args);
 
-      expect(validated.invalidReason, isNotNull);
-      expect(validated.invalidReason!.key, baseError);
-      expect(validated.invalidReason!.value, verboseError);
+      expect(returnedError, error);
     });
 
-    test('returns error when no targer flag is present', () async {
-      const args = ['bump', 'with-path'];
-      const baseError = 'Wrong flag sequence';
-      final verboseError =
-          // ignore: lines_longer_than_80_chars
-          "Command should have at least one of ${validator.targets.take(4).join(', ')} flags";
+    test('returns error when bump and dump are used together', () {
+      const args = ['dump', 'bump'];
+      const error = 'bump and dump flags cannot be used together';
 
-      final validated = await validator.validateArgs(
-        args,
-        isModify: true,
-        userSetPath: false,
-        logger: logger,
-      );
+      final returnedError = validator.checkModifyFlags(args);
 
-      expect(validated.invalidReason, isNotNull);
-      expect(validated.invalidReason!.key, baseError);
-      expect(validated.invalidReason!.value, verboseError);
-    });
-  });
-
-  group('no errors', () {
-    test(
-      'returns no errors when valid args are passed with no set path',
-      () async {
-        const validArgs = ['name', 'major', 'bump', 'yaml-version'];
-
-        final validated = await validator.validateArgs(
-          validArgs,
-          userSetPath: false,
-          logger: logger,
-        );
-
-        expect(validated.invalidReason, isNull);
-      },
-    );
-
-    test("warns and removes 'with-path' flag when path is set", () async {
-      const validArgs = ['name', 'major', 'bump', 'yaml-version', 'with-path'];
-
-      final validated = await validator.validateArgs(
-        validArgs,
-        userSetPath: true,
-        logger: logger,
-      );
-
-      verify(
-        () => logger.warn('Duplicate flags were found when path was set'),
-      ).called(1);
-
-      expect(validated.invalidReason, isNull);
+      expect(returnedError, error);
     });
 
-    test(
-      "warns and removes duplicate 'set-path' flag when path is set",
-      () async {
-        const validArgs = ['name', 'major', 'bump', 'yaml-version', 'set-path'];
+    test('returns error when no target flag is passed in', () {
+      const args = ['bump'];
+      final error = """Command should have at least one of ${validator.targets.take(4).join(', ')} flags""";
 
-        final validated = await validator.validateArgs(
-          validArgs,
-          userSetPath: true,
-          logger: logger,
-        );
+      final returnedError = validator.checkModifyFlags(args);
 
-        verify(
-          () => logger.warn(
-            'Duplicate flags were found when path was set',
-          ),
-        ).called(1);
+      expect(returnedError, error);
+    });
 
-        expect(validated.invalidReason, isNull);
-      },
-    );
+    test('returns error when flags are duplicated', () {
+      const args = ['major', 'major', 'name', 'yaml-version'];
+      const error = 'Found repeated flags:\nmajor -> 2\n';
+
+      final returnedError = validator.checkForDuplicates(args);
+
+      expect(returnedError, error);
+    });
   });
 }
